# refactor(play): simplify translate items for playback