--- a/client/components/Item/Card/Card.vue
+++ b/client/components/Item/Card/Card.vue
@@ -61,7 +61,7 @@
           v-if="overlay && isFinePointer()"
           class="card-overlay d-flex justify-center align-center"
         >
-          <play-button fab :items="[item]" />
+          <play-button fab :item="item" />
           <div
             v-if="overlay"
             class="card-lower-buttons d-flex justify-center align-center"