--- a/client/components/Players/TrackList.vue
+++ b/client/components/Players/TrackList.vue
@@ -135,7 +135,7 @@ export default Vue.extend({
     },
     playTracks(track: BaseItemDto): void {
       this.play({
-        items: this.tracks.Items,
+        item: this.item,
         startFromIndex: this.tracks.Items?.indexOf(track),
         initiator: this.item
       });