--- a/client/nuxt.config.ts
+++ b/client/nuxt.config.ts
@@ -173,6 +173,7 @@ const config: NuxtConfig = {
     plugins: [
       '~/plugins/userLibraryPlugin.ts',
       '~/plugins/itemsPlugin.ts',
+      '~/plugins/tvShowsPlugin.ts',
       '~/plugins/playbackPlugin.ts'
     ]
   },