--- a/client/pages/item/_itemId/index.vue
+++ b/client/pages/item/_itemId/index.vue
@@ -32,7 +32,7 @@
               'ml-0': $vuetify.breakpoint.mdAndUp
             }"
           >
-            <play-button class="mr-2" :items="[item]" />
+            <play-button class="mr-2" :item="item" />
             <like-button :item="item" class="mr-2" />
             <mark-played-button :item="item" class="mr-2" />
             <item-menu :item="item" />