--- a/client/store/items.ts
+++ b/client/store/items.ts
@@ -2,12 +2,14 @@ import Vue from 'vue';
 import { BaseItemDto } from '@jellyfin/client-axios';
 import { ActionTree, GetterTree, MutationTree } from 'vuex';
 import map from 'lodash/map';
-import some from 'lodash/some';
 import forEach from 'lodash/forEach';
+import keyBy from 'lodash/keyBy';
+import merge from 'lodash/merge';
+import union from 'lodash/union';
 
 export interface ItemsState {
   byId: Record<string, BaseItemDto>;
-  allIds: string[];
+  allIds: readonly string[];
 }
 
 export const defaultState = (): ItemsState => ({
@@ -31,14 +33,26 @@ export const getters: GetterTree<ItemsState, ItemsState> = {
 export const mutations: MutationTree<ItemsState> = {
   ADD_ITEM(state: ItemsState, { item }: { item: BaseItemDto }) {
     if (!item.Id) {
-      throw new Error("No item ID in provided item, can't store it");
+      throw new Error('No item ID provided');
     }
+
+    const newAllIds = Array.from(state.allIds);
 
     Vue.set(state.byId, item.Id, item);
 
-    if (!state.allIds.includes(item.Id)) {
-      state.allIds.push(item.Id);
-    }
+    newAllIds.push();
+    state.allIds = Object.freeze(newAllIds);
+  },
+  ADD_ITEMS(state: ItemsState, { items }: { items: BaseItemDto }) {
+    let newById = Object.assign({}, state.byId);
+    let newAllIds = Array.from(state.allIds);
+
+    newById = merge(newById, items);
+
+    state.byId = newById;
+
+    newAllIds = union(newAllIds, Object.keys(items));
+    state.allIds = Object.freeze(newAllIds);
   },
   DELETE_ITEM(state: ItemsState, { id }: { id: string }) {
     delete state.byId[id];
@@ -46,7 +60,10 @@ export const mutations: MutationTree<ItemsState> = {
     const idx = state.allIds.indexOf(id);
 
     if (idx > -1) {
-      state.allIds.splice(idx, 1);
+      const newAllIds = Array.from(state.allIds);
+
+      newAllIds.splice(idx, 1);
+      state.allIds = Object.freeze(newAllIds);
     }
   },
   CLEAR_STATE(state: ItemsState) {
@@ -59,13 +76,9 @@ export const actions: ActionTree<ItemsState, ItemsState> = {
     commit('ADD_ITEM', { item });
   },
   addItems({ commit }, { items }: { items: BaseItemDto[] }) {
-    if (some(items, (item) => !item.Id)) {
-      throw new Error('Item missing ID');
-    }
+    const mappedItems = keyBy(items, 'Id');
 
-    forEach(items, (item) => {
-      commit('ADD_ITEM', { item });
-    });
+    commit('ADD_ITEMS', { items: mappedItems });
   },
   deleteItem({ commit }, { id }: { id: string }) {
     commit('DELETE_ITEM', { id });