* Use the Items store for items in the queue
* Stop trying to be smart and only accept a single item to translate for playback
* Limit the number of fetched items to 300 (This would be 20 hours of continuous music with each song lasting 4 minutes. It mirrors what jf-web does currently.)
* Freeze objects that contain large amounts of data in store/items and store/playbackManager

Benchmark, shuffling a Music library with 319 albums:
```
// Items store
ADD_ITEMS: 105ms
addItems: 110ms

// Playback manager
Translating items for playback: 431ms
Setting queue: 13ms
Setting item index: 28ms
Setting current time: 13ms
Setting initiator and mode: 0ms
Starting playback: 35ms
```