@neopc10 2021-04-15
I can't shuffle playlists with this branch at the moment. It keeps playing the first song. Can we not have this limit in place for the playlists? I run rather long ones that can be up to 1500 tracks and getting only 300 of those is not the least bit desirable 

That is, unless it's fetching 300 items over and over again until all the songs on the playlist are queued.