@MrTimscampi 2021-04-15
> I can't shuffle playlists with this branch at the moment. It keeps playing the first song.

This is fixed as of bb6dcb9.