@MrTimscampi 2021-04-15
This should be ready for reviews. It essentially brings a 10x to ∞x reduction in the time needed to start playback for multiple items.

Some notes:

* The main variation in time now is server-side. It will mostly depend on how efficient the server is at retrieving the items. Our side takes less than 100ms to process everything, even with the full 300 items.
* Other limits were tested, mainly 500 and 1000. Both felt too long, at ~700ms and 1.2s respectively, for fetching items only. The current limits is pretty much instant and comparable in feel to what various music services provide (This was timed against Spotify Desktop and we're pretty much just as fast).
* The entire test suite for playbackManager has been removed. There are two reasons for this: I don't know how to test sub-modules properly and we would need to mock the entire items store, since the queue is now stored there, with all the other items. I'm expecting to revisit this down the line at some point.
* The items store improvements should bring some performance improvements to the rest of the app, but these aren't measured. Since most of the app doesn't implement it yet, we'll likely never truly know, but measures went from around 10s to ~100ms for 300 items)