@ferferga 2021-04-27
As I proposed in Matrix, I think that **loading a single item first and doing the subsequent items in batches** is the best idea for a fast startup + unlimited queues.

I did a quick proof of concept on how my system works **but it can probably be improved**, feedback is welcome.

```
BEFORE (this branch): 
translateItemsForPlayback: 573.175048828125 ms  <== Equals to the time that takes for playback to start

AFTER (no limits):
translateItemsForPlayback: 50.8662109375 ms
Playback started at:  54.82000000483822 ms
translateItemsForPlayback: 2266.677001953125 ms
Queue fully populated at:  2282.3800000041956
Full time:  2337.200000009034 ms

AFTER (300 items only, as this branch):
translateItemsForPlayback: 52.72998046875 ms
Playback started at:  57.09000000206288 ms
translateItemsForPlayback: 1051.066162109375 ms
Queue fully populated at:  1056.205000000773 ms
Full time:  1113.2950000028359 ms
```

As you can see, the batched method has longer total time with 300 items however, the sole reason it starts rightaway after the click action in the button makes the experience even faster by starting really fast.

Source code for the tests are live in https://github.com/jellyfin/jellyfin-vue/tree/simpler-translate-for-playback-tests 
All the pushes in this branch are just to fix merge conflicts with master