@neopc10 2021-05-13
The shuffle button in the media controls appears not to do anything.